requirements.txt
argcomplete==1.10.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.8.0
bibtexparser==1.2.0
bleach==3.2.1
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
file-magic==0.4.0
future==0.18.2
IMAPClient==2.1.0
importlib-metadata==2.0.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
lxml==4.5.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
-e git+git@github.com:DrDub/PAPER.git@e14b1d78b8efa779ba666f5799725f572e2b29b1#egg=paperapp_DrDub
parso==0.7.1
pdfminer.six==20181108
pdftotext==2.1.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.7.1
pypandoc==1.5
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
terminado==0.9.1
testpath==0.4.4
textract==1.6.3
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==1.5.1
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.3.6
zipp==3.2.0
